+++ packages/gpl-core/src/profileMetadata.ts
@@ -26,7 +26,6 @@ export class ProfileMetadata {
    * @deprecated This function is slow and may cause performance issues. Consider using getProfileMetadataByUser instead.
    */
   public async getProfileMetadataAccountsByUser(user: anchor.web3.PublicKey): Promise<anchor.ProgramAccount<any>[]> {
-    console.warn('Warning: getProfileMetadataAccountsByUser is slow and may cause performance issues. Consider using getProfileMetadataByUser instead.');
     const profiles = await this.sdk.profile.getProfileAccountsByUser(user);
     const profilePDAs = profiles.map((p) => p.publicKey);
     let profileMetadataList = [];