+++ packages/react-sdk/src/hooks/post/usePost.ts
@@ -4,30 +4,30 @@ import { PublicKey } from "@solana/web3.js";
 
 const usePost = (sdk: SDK, postAccount: PublicKey) => {
   const [post, setPost] = useState<any>(null);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<Error | null>(null);
+  const [postLoading, setPostLoading] = useState(false);
+  const [postError, setPostError] = useState<Error | null>(null);
 
   const fetchPost = useCallback(
     async () => {
-      setLoading(true);
-      setError(null);
+      setPostLoading(true);
+      setPostError(null);
 
       try {
         const data = await sdk.post.get(postAccount);
 
         setPost(data);
       } catch (err: any) {
-        setError(err);
+        setPostError(err);
       } finally {
-        setLoading(false);
+        setPostLoading(false);
       }
     }, [sdk, postAccount]);
 
   useEffect(() => {
   fetchPost();
-  }, [fetchPost]);
+  }, []);
 
-  return { post, loading, error };
+  return { post, postLoading, postError };
 };
 
 export { usePost };