+++ packages/react-sdk/src/hooks/user/useCreateUser.ts
@@ -4,36 +4,52 @@ import { PublicKey } from "@solana/web3.js";
 
 const useCreateUser = (sdk: SDK) => {
   const [userPDA, setUserPDA] = useState<PublicKey | null>(null);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<Error | null>(null);
+  const [isCreatingUser, setIsCreatingUser] = useState(false);
+  const [createUserError, setCreateUserError] = useState<Error | null>(null);
 
   const create = useCallback(
     async (owner: PublicKey) => {
-      setLoading(true);
-      setError(null);
+      setIsCreatingUser(true);
+      setCreateUserError(null);
 
       try {
         const data = await createUserIxMethodBuilder(owner);
         await data?.rpc();
       } catch (err: any) {
-        setError(err);
+        setCreateUserError(err);
       } finally {
-        setLoading(false);
+        setIsCreatingUser(false);
       }
     },
     [sdk]
   );
 
+  const getOrCreate = useCallback(
+    async (owner: PublicKey) => {
+      setIsCreatingUser(true);
+      setCreateUserError(null);
+
+      try {
+        const userPDA = await sdk.user.getOrCreate(owner);
+        setUserPDA(userPDA);
+      } catch (err: any) {
+        setCreateUserError(err);
+      } finally {
+        setIsCreatingUser(false);
+      }
+    },
+    [sdk]);
+
   const createUserIxMethodBuilder = useCallback(
     async (owner: PublicKey) => {
-      setError(null);
+      setCreateUserError(null);
 
       try {
         const user = await sdk.user.create(owner);
         setUserPDA(user?.userPDA);
         return user.instructionMethodBuilder;
       } catch (err: any) {
-        setError(err);
+        setCreateUserError(err);
         return null;
       }
     },
@@ -42,10 +58,11 @@ const useCreateUser = (sdk: SDK) => {
 
   return {
     create,
+    getOrCreate,
     createUserIxMethodBuilder,
     userPDA,
-    loading,
-    error
+    isCreatingUser,
+    createUserError
   };
 };