+++ packages/react-sdk/CHANGELOG.md
@@ -1,7 +1,23 @@
-# Change Log
+## [1.0.8](https://github.com/gumhq/sdk/compare/v1.0.10...v1.0.8) (2023-03-01)
 
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+### Bug Fixes
+
+- **react-sdk hooks:** remove unnecessary dependency in hooks ([214a595](https://github.com/gumhq/sdk/commit/214a595f50ab35d129c662cd20e495801d5329f9))
+
+### Features
+
+- **@gumhq/react-sdk:** export feed hooks ([cafaef3](https://github.com/gumhq/sdk/commit/cafaef3cd255b38819d805950f918a434afbc1be))
+
+## [1.0.7](https://github.com/gumhq/sdk/compare/v1.0.10...v1.0.7) (2023-02-23)
+
+### Bug Fixes
+
+- **@gumhq-react-sdk:** usage code in readme ([41929b8](https://github.com/gumhq/sdk/commit/41929b8af6ad0bbf0ffb2ec0b39777bd93cabc29))
+
+### Features
+
+- **@gumhq/react-sdk:** add Feed hooks ([92c0c8a](https://github.com/gumhq/sdk/commit/92c0c8ade1508b6445d7a0fb377def4067c8ab81))
+- **@gumhq/react-sdk:** added getOrCreate() method for user and profile in react sdk hook ([7f943bb](https://github.com/gumhq/sdk/commit/7f943bb1542996162c6b1cd47351f0b64f92b098))
 
 ## [1.0.7-alpha.0](https://github.com/gumhq/sdk/compare/@gumhq/react-sdk@1.0.6...@gumhq/react-sdk@1.0.7-alpha.0) (2023-02-17)